Moved KeyboardWatcherActivity to separate module.
