+++ app/build.gradle
@@ -53,12 +53,11 @@ dependencies {
 
     implementation libraries.placeHolderView
 
-    implementation libraries.keyboardWatcher
-
     testImplementation libraries.junit
     testImplementation libraries.mockito
     testImplementation libraries.robolectric
     implementation project(path: ':curvedbottombar')
     implementation project(path: ':viewpagerwithoutfragment')
     implementation project(path: ':fabanimation')
+    implementation project(path: ':keyboardwatcher')
 }
