+++ app/src/main/java/com/example/kirchhoff/example/MainActivity.java
@@ -10,7 +10,6 @@ import com.example.kirchhoff.example.ui.drawer.simple.SimpleActivity;
 import com.example.kirchhoff.example.ui.other.CoordinatorBehaviorActivity;
 import com.example.kirchhoff.example.ui.other.CountdownLabelActivity;
 import com.example.kirchhoff.example.ui.other.CustomActionBarActivity;
-import com.example.kirchhoff.example.ui.other.KeyboardWatcherActivity;
 import com.example.kirchhoff.example.ui.recycler.diffutil.ex1.DiffUtilActivity;
 import com.example.kirchhoff.example.ui.recycler.diffutil.ex2.ThingDiffUtilActivity;
 import com.example.kirchhoff.example.ui.recycler.fast.FastRecyclerViewActivity;
@@ -24,6 +23,7 @@ import com.example.kirchhoff.example.ui.tinder.TinderActivity;
 import com.example.kirchhoff.example.ui.view.ViewActivity;
 import com.kirchhoff.curvedbottombar.CurvedBottomBarActivity;
 import com.kirchhoff.fabanimation.FabAnimationActivity;
+import com.kirchhoff.keyboardwatcher.KeyboardWatcherActivity;
 import com.kirchhoff.viewpagerwithoutfragment.VPWithoutFragmentActivity;
 
 public class MainActivity extends AppCompatActivity implements View.OnClickListener {
@@ -101,7 +101,8 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
     }
 
     public void keyboardWatcherExample(View view) {
-        KeyboardWatcherActivity.start(this);
+        Intent intent = new Intent(this, KeyboardWatcherActivity.class);
+        startActivity(intent);
     }
 
     public void fastRecyclerViewExample(View view) {
