+++ keyboardwatcher/src/main/AndroidManifest.xml
@@ -1,0 +1,11 @@
+<manifest package="com.kirchhoff.keyboardwatcher"
+	xmlns:android="http://schemas.android.com/apk/res/android">
+
+	<application
+		android:allowBackup="true"
+		android:icon="@mipmap/ic_launcher"
+		android:label="@string/app_name"
+		android:roundIcon="@mipmap/ic_launcher_round"
+		android:supportsRtl="true"
+		android:theme="@style/AppTheme" />
+</manifest>
