+++ keyboardwatcher/src/main/res/layout/a_keyboard_watcher.xml
@@ -4,7 +4,6 @@
     android:layout_height="match_parent"
     android:orientation="vertical">
 
-
     <EditText
         android:id="@+id/etFirst"
         android:layout_width="match_parent"
